[WIP] move mdb, make isfdb just db
