--- data_base/isf_data_base/__init__.py
@@ -1,25 +1,0 @@
-# In Silico Framework
-# Copyright (C) 2025  Max Planck Institute for Neurobiology of Behavior - CAESAR
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-# The full license text is also available in the LICENSE file in the root of this repository.
-"""
-Data base system using JSON format for metadata.
-
-See also:
-    :py:class:`~data_base.isf_data_base.ISFDataBase` for the main data base class, and
-    the module :py:mod:`data_base.isf_data_base.IO` provides readers and writers for various file formats.
-"""
-
-from .isf_data_base import ISFDataBase
