--- data_base/model_data_base/IO/LoaderDumper/dask_to_msgpack.py
@@ -1,31 +1,0 @@
-# In Silico Framework
-# Copyright (C) 2025  Max Planck Institute for Neurobiology of Behavior - CAESAR
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-# The full license text is also available in the LICENSE file in the root of this repository.
-from . import dask_to_categorized_msgpack
-import os
-
-Loader = dask_to_categorized_msgpack.Loader
-check = dask_to_categorized_msgpack.check
-
-
-def dump(obj, savedir, repartition=False, scheduler=None, client=None):
-    import os
-    return dask_to_categorized_msgpack.dump(obj,
-                                            savedir,
-                                            repartition=repartition,
-                                            scheduler=scheduler,
-                                            categorize=False,
-                                            client=client)
